+++ src/Firebase/Exception/Auth/OperationNotAllowed.php
@@ -7,7 +7,7 @@ use Throwable;
 
 class OperationNotAllowed extends AuthException
 {
-    const IDENTIFER = 'OPERATION_NOT_ALLOWED';
+    public const IDENTIFER = 'OPERATION_NOT_ALLOWED';
 
     public function __construct($code = 0, Throwable $previous = null)
     {