+++ src/Firebase/Messaging/MessageTarget.php
@@ -8,11 +8,11 @@ use Kreait\Firebase\Exception\InvalidArgumentException;
 
 final class MessageTarget
 {
-    const CONDITION = 'condition';
-    const TOKEN = 'token';
-    const TOPIC = 'topic';
+    public const CONDITION = 'condition';
+    public const TOKEN = 'token';
+    public const TOPIC = 'topic';
 
-    const TYPES = [
+    public const TYPES = [
         self::CONDITION, self::TOKEN, self::TOPIC,
     ];